refactor: modify codebase to use db-subgraph plugin
