+++ package.json
@@ -10,13 +10,12 @@
     "generateDocs": "typedoc",
     "lint:fix": "eslint --fix .",
     "lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
-    "postinstall": "pnpm wagmi-generate && pnpm subgraph-codegen",
+    "postinstall": "pnpm wagmi-generate",
     "prepare": "husky",
     "prettier:fix": "prettier --write --ignore-unknown .",
     "preview": "vite preview",
     "routes:generate": "tsr generate",
     "routes:watch": "tsr watch",
-    "subgraph-codegen": "graphql-codegen --config ./src/subgraphs/codegen.ts",
     "test:coverage": "vitest run --coverage",
     "test:watch": "vitest",
     "test": "vitest run",
@@ -24,7 +23,6 @@
   },
   "dependencies": {
     "@faceless-ui/modal": "3.0.0-beta.2",
-    "@graphql-typed-document-node/core": "^3.2.0",
     "@lifi/sdk": "^3.0.1",
     "@t3-oss/env-core": "^0.10.1",
     "@tanstack/react-query": "^5.36.0",
@@ -33,6 +31,7 @@
     "@uniswap/default-token-list": "^11.19.0",
     "@vercel/analytics": "^1.3.1",
     "connectkit": "^1.8.2",
+    "db-subgraph": "^0.1.0",
     "db-ui-toolkit": "github:BootNodeDev/db-ui-toolkit#0.3.7",
     "graphql": "^16.9.0",
     "graphql-request": "^7.1.0",
@@ -46,7 +45,7 @@
     "react-number-format": "^5.4.0",
     "styled-components": "^6.1.12",
     "use-debounce": "^10.0.1",
-    "viem": "~2.10.9",
+    "viem": "^2.21.1",
     "wagmi": "^2.9.2",
     "zod": "^3.23.8"
   },
@@ -54,7 +53,7 @@
     "@commitlint/cli": "^19.3.0",
     "@commitlint/config-conventional": "^19.2.2",
     "@graphql-codegen/cli": "^5.0.2",
-    "@graphql-codegen/client-preset": "^4.3.2",
+    "@graphql-typed-document-node/core": "^3.2.0",
     "@parcel/watcher": "^2.4.1",
     "@tanstack/react-query-devtools": "^5.36.0",
     "@tanstack/router-cli": "^1.32.2",
